[SAC-124][SQL] Hive Warehouse Connector support (#125)
This PR proposes to add HWC support in SAC.

Hive Warehouse Connector currently supports four types of operations:
  1. SQL / DataFrame Read (batch read)
  2. SQL / DataFrame Write (batch write)
  3. SQL / DataFrame Write in streaming manner (batch write in streaming)
  4. Structured Streaming Write (streaming write)

For 1., it is supported by looking logical plans (if available) or physical
plans up at `HWCEntities` for every execution plan being processed above
when it's possible.

For 2. and 3., it checks only when the execution plan is `WriteToDataSourceV2Exec`.
It checks the write implementation of DataSourceV2 is HWC or not and dispatches to harvest
appropriate entities.

For 4., it is same as 2. and 3. but it reuses Kafka harvester to handle input sources
under the hood when it dispatches to harvest.

Manually tested:

![screen shot 2018-11-20 at 4 56 49 pm](https://user-images.githubusercontent.com/6477701/48814851-65d2b900-ed77-11e8-8a43-a9f50cd58443.png)

The examples at https://community.hortonworks.com/content/kbentry/223626/integrating-apache-hive-with-apache-spark-hive-war.html were used and tested.

Closes #124