[SAC-141] Make SAC recognise HWC's SQL API (#142)
This PR proposes to make HWC's SQL syntax recognisable. Currently SAC and HWC only know the query that has to be pushed down to Hive side so here we don't know which tables are to be proceeded. To work around, we use Spark's parser driver to identify tables. Once we know the table identifiers by this, it looks up Hive entities to find Hive tables out.

Manually tested via a Yarn secure cluster.

![screen shot 2018-12-04 at 10 33 01 am](https://user-images.githubusercontent.com/6477701/49415466-d0d9b200-f7b1-11e8-9d53-9290746bd03c.png)

Closes #141