[SAC-154][SQL] Create only table and columns entities (if enabled) for alter table event (#155)
This patch proposes to separate handling of `create table` and `alter table`, and avoid creating entities for database and storagedesc for the event of `alter table`.

Modified UT to ensure database and storagedesc entities are not created during alter table event.
Also manually tested with Atlas 1.1 & Spark 2.4 to ensure there's no audit update on both database and storagedesc entities.

* table entity
![screen shot 2018-12-12 at 1 38 12 pm](https://user-images.githubusercontent.com/1317309/49848118-f18bb280-fe16-11e8-9768-8891b23d0343.png)

* db entity
![screen shot 2018-12-12 at 1 38 01 pm](https://user-images.githubusercontent.com/1317309/49848124-f8b2c080-fe16-11e8-9d56-b6d5f37e6bd7.png)

* storage entity
![screen shot 2018-12-12 at 1 37 46 pm](https://user-images.githubusercontent.com/1317309/49848135-0405ec00-fe17-11e8-8e9f-bc41fd35e307.png)

Please look for create/update on `Wed Dec 12 2018 13:37:09`. Audit is only added to table.

This closes #154