[SAC-156] Spark 2.3 - Support custom atlas cluster name for Kafka topics (source/sink) (#158)
## What changes were proposed in this pull request?

This patch proposes applying #103 against Spark 2.3 as well. Previously we dropped applying Spark 2.3 because it requires custom patch for Spark 2.3, and now we just decided to proceed with custom patch. 

This patch leverages reflection to access newly exposed field. If Spark is not patched, SAC will print out WARN message and will not extract custom atlas cluster name.

## How was this patch tested?

The change of this patch is verified from #157 which is only used when dealing with Spark 2.3. In Spark 2.4 it will not be used.

This closes #156