[SAC-192][SQL] Support JDBC (RDBMS) based operations (#200)
## What changes were proposed in this pull request?

The SAC plugin was unable to track JDBC based operations as it was unable to identify the output, as indicated [in the following issue](https://github.com/hortonworks-spark/spark-atlas-connector/issues/192). The changes added the logic to identify a JDBC based operation and handle it. Minor refactoring was also done. And some unit tests were added to increase code coverage. Cobertura plugin was used to identify what unit tests needed to be added.

## How was this patch tested?

Unit tests were added, as well as manual testing

Below screenshot shows an example where a file was uploaded to a MySQL table and the contents copied to another table.

![screenshot from 2019-01-28 15-40-09](https://user-images.githubusercontent.com/1842813/51843390-24303280-2313-11e9-87d2-b64f22c5bc2c.png)

This closes #200