[SAC-252][CORE] Skip evaluating streaming event when query is not physically executed yet (#254)
## What changes were proposed in this pull request?

Please refer #252 to see the rationalization of this patch.

This patch fixes the intermittent NPE occurred in SAC side when executing Streaming query, which is occurred due to receiving streaming event before executing actual batch. This can be easily reproduced via executing streaming query and let query consuming all the available input data, and stop the query, and restart the query in new spark session (spark-shell) which leverage same checkpoint.

This patch checks whether the last executed physical plan is available, and skips processing it if the plan is not available. This effectively makes SAC to wait until the plan is available (since streaming event will be sent at least per each batch, so eventually it will be available).

## How was this patch tested?

Manually tested: patched SAC didn't throw NPE with no batch but just waited for new event, and when I provided new records to one of topic, SAC picked the event and created a new lineage.
 
<img width="1058" alt="Screen Shot 2019-05-27 at 5 23 07 PM" src="https://user-images.githubusercontent.com/1317309/58406241-2083a100-80a4-11e9-97a5-d5adb0720f05.png">

NOTE: there're actually three topics in source side, but only one source is captured - it's a known-issue and I filed #253 to track this.

This closes #252